% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_censor_marks}
\alias{add_censor_marks}
\title{Add censor marks to KM plot}
\usage{
add_censor_marks(p, df, censor.size = 0.5, censor.stroke = 1.5, shape = "|")
}
\arguments{
\item{p}{ggplot object}

\item{df}{Plotting dataframe}

\item{censor.size}{Size of censor marks}

\item{censor.stroke}{Stroke of censor marks}

\item{shape}{Shape for censor marks, defaults to "|", but can use an
character or standard geom_point shapes (0-24)}
}
\description{
Add censor marks to KM plot
}
