% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_confidence_bands}
\alias{add_confidence_bands}
\title{Add confidence bands to plot}
\usage{
add_confidence_bands(p, df, type, event = "col", plot.event = 1)
}
\arguments{
\item{p}{ggplot object}

\item{df}{Plotting dataframe}

\item{type}{Plot type}

\item{event}{Event distinction method}

\item{plot.event}{Events to plot}
}
\description{
Add confidence bands to plot
}
