% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_km_hazard_ratios}
\alias{add_km_hazard_ratios}
\title{Add hazard ratios to strata labels}
\usage{
add_km_hazard_ratios(
  stratalabs,
  data,
  response,
  cov,
  type,
  plot.event = 1,
  HR = FALSE,
  HR_pval = FALSE,
  HR.digits = 2,
  HR.pval.digits = 3
)
}
\arguments{
\item{stratalabs}{Original strata labels}

\item{data}{Input data}

\item{response}{Time and status variables}

\item{cov}{Covariate}

\item{type}{Model type ("KM" or "CIF")}

\item{plot.event}{Event for CIF}

\item{HR}{Whether to include HR}

\item{HR_pval}{Whether to include HR p-value}

\item{HR.digits}{Number of digits for HR}

\item{HR.pval.digits}{Number of digits for HR p-value}
}
\description{
Add hazard ratios to strata labels
}
