% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_median_text}
\alias{add_median_text}
\title{Add median survival text to plot}
\usage{
add_median_text(
  p,
  type,
  multiple_lines,
  median_txt,
  median.pos = NULL,
  times,
  ylim,
  median.size = 3,
  plot.event = 1,
  eventlabs = NULL
)
}
\arguments{
\item{p}{ggplot object}

\item{type}{Plot type}

\item{multiple_lines}{Whether multiple strata}

\item{median_txt}{Median text}

\item{median.pos}{Position for median text}

\item{times}{Time breaks}

\item{ylim}{Y-axis limits}

\item{median.size}{Text size}

\item{plot.event}{Events being plotted}

\item{eventlabs}{Event labels}
}
\description{
Add median survival text to plot
}
