% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_set_time_lines}
\alias{add_set_time_lines}
\title{Add survival at set time lines to plot}
\usage{
add_set_time_lines(p, set.surv, set.lsize = 1)
}
\arguments{
\item{p}{ggplot object}

\item{set.surv}{Data frame with survival at set times}

\item{set.lsize}{Line size}
}
\description{
Add survival at set time lines to plot
}
