% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_set_time_text}
\alias{add_set_time_text}
\title{Add survival at set time text to plot}
\usage{
add_set_time_text(
  p,
  type,
  multiple_lines,
  set.surv.text,
  set.pos = NULL,
  times,
  ylim,
  set.size = 3,
  plot.event = 1,
  eventlabs = NULL
)
}
\arguments{
\item{p}{ggplot object}

\item{type}{Plot type}

\item{multiple_lines}{Whether multiple strata}

\item{set.surv.text}{Survival text}

\item{set.pos}{Position for survival text}

\item{times}{Time breaks}

\item{ylim}{Y-axis limits}

\item{set.size}{Text size}

\item{plot.event}{Events being plotted}

\item{eventlabs}{Event labels}
}
\description{
Add survival at set time text to plot
}
