% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{add_statistical_tests}
\alias{add_statistical_tests}
\title{Add statistical test results to plot}
\usage{
add_statistical_tests(
  p,
  type,
  multiple_lines,
  pval_result,
  pval.pos = NULL,
  times,
  xlim,
  ylim,
  psize = 3.5,
  pval.digits = 3,
  plot.event = 1,
  eventlabs = NULL
)
}
\arguments{
\item{p}{ggplot object}

\item{type}{Plot type}

\item{multiple_lines}{Whether multiple strata}

\item{pval_result}{P-value}

\item{pval.pos}{Position for p-value text}

\item{times}{Time breaks}

\item{xlim}{X-axis limits}

\item{ylim}{Y-axis limits}

\item{psize}{Text size for p-value}

\item{pval.digits}{Number of digits for p-value}

\item{plot.event}{Events being plotted}

\item{eventlabs}{Event labels}
}
\description{
Add statistical test results to plot
}
