% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{apply_scales_and_guides}
\alias{apply_scales_and_guides}
\title{Apply colour and linetype scales}
\usage{
apply_scales_and_guides(
  p,
  col,
  linetype = NULL,
  stratalabs,
  eventlabs = NULL,
  multiple_lines,
  plot.event = 1,
  event = "col"
)
}
\arguments{
\item{p}{ggplot object}

\item{col}{colours vector}

\item{linetype}{Line types vector}

\item{stratalabs}{Strata labels}

\item{eventlabs}{Event labels}

\item{multiple_lines}{Whether multiple strata}

\item{plot.event}{Events being plotted}

\item{event}{How events are distinguished}
}
\description{
Apply colour and linetype scales
}
