% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_arrow_segments}
\alias{build_arrow_segments}
\title{Build arrow segments for CIs extending beyond xlim}
\usage{
build_arrow_segments(tab, xlim, use_linetype = FALSE)
}
\arguments{
\item{tab}{forest plot data with ci_low_arrow and ci_high_arrow columns}

\item{xlim}{numeric vector of length 2}

\item{use_linetype}{logical, include linetype aesthetic?}
}
\value{
list of ggplot layers
}
\description{
Returns a list of ggplot layers for left and right arrows.
}
\keyword{internal}
