% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_forest_ggplot}
\alias{build_forest_ggplot}
\title{Build a forest plot from prepared data}
\usage{
build_forest_ggplot(
  tab,
  x_lab,
  show_linetype = FALSE,
  colours = "default",
  showEst = TRUE,
  showRef = TRUE,
  logScale = TRUE,
  nxTicks = 5,
  showN = TRUE,
  showEvent = TRUE,
  xlim = c(0.2, 5)
)
}
\arguments{
\item{tab}{data frame prepared by prepare_forest_data and order_forest_data}

\item{x_lab}{character label for the x-axis}

\item{show_linetype}{logical, should adjusted/unadjusted be distinguished by
linetype and shape? TRUE when showing both adjusted and unadjusted.}

\item{colours}{colour specification: "default" or a vector of 3 colours for
risks less than, equal to, and greater than 1.0}

\item{showEst}{logical, should estimates be shown in labels?}

\item{showRef}{logical, should reference levels be shown?}

\item{logScale}{logical, should x-axis be log-scaled?}

\item{nxTicks}{number of x-axis tick marks}

\item{showN}{logical, show N on secondary axis?}

\item{showEvent}{logical, show events on secondary axis?}

\item{xlim}{numeric vector of length 2 for x-axis limits, or NULL}
}
\description{
Internal function that handles all ggplot construction for forest plots.
Called by forestplotMV and forestplotUV after data preparation.
}
\keyword{internal}
