% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_log_scale}
\alias{build_log_scale}
\title{Build log scale for forest plot x-axis}
\usage{
build_log_scale(tab, xlim, nxTicks)
}
\arguments{
\item{tab}{forest plot data}

\item{xlim}{numeric vector of length 2, or NULL}

\item{nxTicks}{number of tick marks}
}
\value{
a scale_x_log10 layer
}
\description{
Build log scale for forest plot x-axis
}
\keyword{internal}
