% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_secondary_axis}
\alias{build_secondary_axis}
\title{Build secondary axis specification}
\usage{
build_secondary_axis(tab, yLabels, showN, showEvent)
}
\arguments{
\item{tab}{forest plot data}

\item{yLabels}{data frame with y.pos and labels}

\item{showN}{logical}

\item{showEvent}{logical}
}
\value{
list with axis (scale_y_continuous) and theme elements
}
\description{
Build secondary axis specification
}
\keyword{internal}
