% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_refactored_compactsum.r
\name{calc_CramerV}
\alias{calc_CramerV}
\title{Calculate Cramer's V with bootstrap CI}
\usage{
calc_CramerV(chisq_test, CIwidth = 0.95)
}
\description{
Calculate Cramer's V with bootstrap CI
}
\keyword{internal}
