% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_refactored_compactsum.r
\name{calc_omegaSq}
\alias{calc_omegaSq}
\title{Calculate Omega squared with exact CI from eta squared, or bootstrap CI}
\usage{
calc_omegaSq(anova_test, CIwidth = 0.95)
}
\description{
Calculate Omega squared with exact CI from eta squared, or bootstrap CI
}
\keyword{internal}
