% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{calculate_and_add_time_specific_estimates}
\alias{calculate_and_add_time_specific_estimates}
\title{Calculate survival/CIF at specific time points and add to labels}
\usage{
calculate_and_add_time_specific_estimates(
  sfit = NULL,
  fit = NULL,
  stratalabs,
  type = "KM",
  plot.event = 1,
  set.time.text = NULL,
  set.time = NULL,
  set.time.line = FALSE,
  set.time.CI = FALSE,
  set.time.digits = 3
)
}
\arguments{
\item{sfit}{Survival fit object (for KM)}

\item{fit}{CIF fit object}

\item{stratalabs}{Strata labels}

\item{type}{Model type}

\item{plot.event}{Events to plot (for CIF)}

\item{set.time.text}{Text label for time points}

\item{set.time}{Time points to evaluate}

\item{set.time.line}{boolean to specify if you want the survival added as
lines to the plot at a specified point}

\item{set.time.CI}{Whether to include CI}

\item{set.time.digits}{Number of digits}
}
\description{
Calculate survival/CIF at specific time points and add to labels
}
