% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{calculate_cif_timepoints}
\alias{calculate_cif_timepoints}
\title{Calculate CIF Estimates at Specific Time Points}
\usage{
calculate_cif_timepoints(
  fit,
  plot.event,
  set.time,
  set.time.CI = FALSE,
  set.time.digits = 3,
  multiple_lines = FALSE
)
}
\arguments{
\item{fit}{CIF fit object}

\item{plot.event}{Events to plot}

\item{set.time}{Time points to evaluate}

\item{set.time.CI}{Whether to include confidence intervals}

\item{set.time.digits}{Number of digits}

\item{multiple_lines}{Whether there are multiple strata}
}
\value{
Data frame with time-specific estimates
}
\description{
Calculate CIF Estimates at Specific Time Points
}
