% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{calculate_median_times}
\alias{calculate_median_times}
\title{Calculate median survival times}
\usage{
calculate_median_times(sfit = NULL, type = "KM", fit = NULL, plot.event = 1)
}
\arguments{
\item{sfit}{Survival fit object}

\item{type}{Model type ("KM" or "CIF")}

\item{fit}{CIF fit object (for CIF type)}

\item{plot.event}{Events to plot (for CIF)}
}
\description{
Calculate median survival times
}
