% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{calculate_time_specific_estimates}
\alias{calculate_time_specific_estimates}
\title{Calculate survival/CIF at specific time points}
\usage{
calculate_time_specific_estimates(
  sfit = NULL,
  fit = NULL,
  times,
  type = "KM",
  plot.event = 1
)
}
\arguments{
\item{sfit}{Survival fit object (for KM)}

\item{fit}{CIF fit object}

\item{times}{Time points to evaluate}

\item{type}{Model type}

\item{plot.event}{Events to plot (for CIF)}
}
\description{
Calculate survival/CIF at specific time points
}
