% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_refactored_compactsum.r
\name{categ_xvar_helper}
\alias{categ_xvar_helper}
\title{Helper for categorical variables by group}
\usage{
categ_xvar_helper(
  xvar_level,
  data,
  xvar,
  grp,
  digits.cat = 0,
  percentage = "col",
  pct_symbol = TRUE
)
}
\description{
Helper for categorical variables by group
}
\keyword{internal}
