% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblCode.R
\name{clear_labels}
\alias{clear_labels}
\title{Clear variable labels from a data frame}
\usage{
clear_labels(data)
}
\arguments{
\item{data}{the data frame to remove labels from}
}
\description{
This function will remove all label attributes from variables in the data.
}
\details{
To change or remove individual labels use set_labels or set_var_labels
}
\examples{
# Set a few variable labels for ctDNA
data("ctDNA")
ctDNA <- ctDNA |> set_var_labels(
   ctdna_status="detectable ctDNA",
  cohort="A cohort label")
# Clear all variable data frames and check
clear_labels(ctDNA)
}
