% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{create_base_plot}
\alias{create_base_plot}
\title{Create base ggplot for survival curves}
\usage{
create_base_plot(
  df,
  type,
  xlab = "Time",
  ylab = "Survival Probability",
  multiple_lines,
  plot.event = 1,
  event = "col",
  lsize = 0.5,
  fsize,
  col,
  legend.pos = "right"
)
}
\arguments{
\item{df}{Plotting dataframe}

\item{type}{Plot type ("KM" or "CIF")}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{multiple_lines}{Whether multiple strata}

\item{plot.event}{Events to plot}

\item{event}{How to distinguish events ("col" or "linetype")}

\item{lsize}{Line size}

\item{fsize}{Font size}

\item{col}{vector of colours}

\item{legend.pos}{Legend position, can be "top", right" or "none"}
}
\description{
Create base ggplot for survival curves
}
