% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{create_cif_dataframe}
\alias{create_cif_dataframe}
\title{Create CIF Plotting Data Frame}
\usage{
create_cif_dataframe(
  fit,
  gsep,
  plot.event,
  stratalabs,
  conf.type = "log",
  flip.CIF = FALSE,
  eventlabs = NULL,
  cov = NULL,
  data = NULL
)
}
\arguments{
\item{fit}{CIF fit object}

\item{gsep}{Group separator from fit_cif_model}

\item{plot.event}{Events to plot}

\item{stratalabs}{Strata labels}

\item{conf.type}{Confidence interval type}

\item{flip.CIF}{Whether to flip the CIF curve}

\item{eventlabs}{Event labels}

\item{cov}{Covariate for proper factor levels}

\item{data}{Original data (for factor levels)}
}
\value{
Data frame for plotting
}
\description{
Create CIF Plotting Data Frame
}
