% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{create_cif_risk_table_sfit}
\alias{create_cif_risk_table_sfit}
\title{Create Survival Fit for Risk Table in CIF}
\usage{
create_cif_risk_table_sfit(data, response, cov = NULL)
}
\arguments{
\item{data}{Input data}

\item{response}{Time and status variables}

\item{cov}{Covariate (optional)}
}
\value{
Survival fit object for risk table
}
\description{
Create Survival Fit for Risk Table in CIF
}
