% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{create_km_dataframe}
\alias{create_km_dataframe}
\title{Create plotting dataframe for KM curves}
\usage{
create_km_dataframe(sfit, stratalabs, conf.curves = FALSE, conf.type = "log")
}
\arguments{
\item{sfit}{Survival fit object}

\item{stratalabs}{Strata labels}

\item{conf.curves}{Whether to include confidence intervals}

\item{conf.type}{Confidence interval type}
}
\description{
Create plotting dataframe for KM curves
}
