% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{create_risk_table}
\alias{create_risk_table}
\title{Create risk table for survival plot}
\usage{
create_risk_table(
  sfit,
  times,
  stratalabs,
  stratalabs.table = NULL,
  Numbers_at_risk_text = "At Risk",
  multiple_lines = TRUE,
  col = NULL,
  fsize = 12,
  nsize = 3
)
}
\arguments{
\item{sfit}{Survival fit object}

\item{times}{Time points for risk table}

\item{stratalabs}{Strata labels}

\item{stratalabs.table}{Table-specific strata labels}

\item{multiple_lines}{Whether multiple strata}

\item{col}{Colors for strata}

\item{fsize}{Font size}

\item{nsize}{Number size in table}
}
\description{
Create risk table for survival plot
}
