% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctDNA}
\alias{ctDNA}
\title{Tumour size change over time

Longitudinal changes in tumour size since baseline for patients by changes in ctDNA status (clearance, decrease or increase) since baseline.}
\format{
A data frame with 270 rows and 5 variables:
\describe{
\item{id}{Patient ID}
\item{cohort}{Study Cohort}
\item{ctdna_status}{Change in ctDNA since baseline}
\item{time}{Number of weeks on treatment}
\item{size_change}{Percentage change in tumour measurement}
}
}
\source{
\url{https://www.nature.com/articles/s43018-020-0096-5}
}
\usage{
data('ctDNA')
}
\description{
Tumour size change over time

Longitudinal changes in tumour size since baseline for patients by changes in ctDNA status (clearance, decrease or increase) since baseline.
}
\keyword{datasets}
