% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{.extract_ggplot_colours}
\alias{.extract_ggplot_colours}
\title{Function to extract ggplot colours}
\usage{
.extract_ggplot_colours(p, grp.levels)
}
\description{
Function to extract ggplot colours
}
