% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{.set_large_dash_as_ytext}
\alias{.set_large_dash_as_ytext}
\title{Function to set large dash as y-axis text}
\usage{
.set_large_dash_as_ytext(plot)
}
\description{
Function to set large dash as y-axis text
}
