% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarLevels.R
\name{esc_rx}
\alias{esc_rx}
\title{Escape regular expression metacharacters}
\usage{
esc_rx(x)
}
\arguments{
\item{x}{character vector}
}
\value{
\code{x} with regex metacharacters escaped
}
\description{
Variable names can legally contain regex metacharacters, most commonly the
parentheses in terms such as \code{poly(x, 2)} or \code{log(x)}. Names used to build
patterns must be escaped first.
}
\keyword{internal}
