% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{etsum}
\alias{etsum}
\title{Get event time summary dataframe}
\usage{
etsum(data, response, group = 1, times = c(12, 24))
}
\arguments{
\item{data}{dataframe containing data}

\item{response}{character vector with names of columns to use for response}

\item{group}{string specifying the column name of stratification variable}

\item{times}{numeric vector of times you want survival time probabilities
for.}
}
\description{
This function will output a dataframe with useful summary statistics from a
coxph model
}
\keyword{dataframe}
