% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{excelCol}
\alias{excelCol}
\title{Retrieve columns number from spreadsheet columns specified as unquoted letters}
\usage{
excelCol(...)
}
\arguments{
\item{...}{unquoted excel column headers (i.e. excelCol(A,CG,AA)) separated by commas}
}
\value{
a numeric vector corresponding to columns in a spreadsheet
}
\description{
Retrieve columns number from spreadsheet columns specified as unquoted letters
}
\examples{
## Find the column numbers for excel columns AB, CE and BB
excelCol(AB,CE,bb)
## Get the columns between A and K and Z
excelCol(A-K,Z)
}
