% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{excelColLetters}
\alias{excelColLetters}
\title{Retrieve spreadsheet column letter-names from columns indices}
\usage{
excelColLetters(columnIndices)
}
\arguments{
\item{columnIndices}{vector of integer column indices}
}
\value{
a character vector corresponding to the spreadsheet column headings
}
\description{
Creates a vector of spreadsheet-style letter-names corresponding to column numbers
}
\details{
This is the inverse function of excelCol
}
\examples{
## Find the column numbers for excel columns AB, CE and BB
colIndices <- excelCol(AB,CE,bb)
## Go back to the column names
excelColLetters(colIndices)
}
