% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblCode.R
\name{extractLabels}
\alias{extractLabels}
\title{Extract variable labels from labelled data and set}
\usage{
extractLabels(data, default = FALSE)
}
\arguments{
\item{data}{data frame with labelled variables to be extracted.}

\item{default}{logical, should the variable labels be used as the default
labels for all data frames, default is FALSE.}
}
\description{
Retrieve labels set using the haven, expss or sjlabelled and set the variable
label table.
}
\details{
By default the variable names will be associated with the data frame. To make
the variable labels the default table specify default=TRUE.

If no labels are supplied for a variable then the variable name will be used
with '_' replaced by a space.
}
\examples{
# example code (not run)
# library(sjlabelled)
# data(efc)
# extractLabels(efc)
}
