% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{fit_cif_model}
\alias{fit_cif_model}
\title{Fit Cumulative Incidence Function curves}
\usage{
fit_cif_model(data, response, cov = NULL)
}
\arguments{
\item{data}{Input data}

\item{response}{Time and status variables}

\item{cov}{Covariate (optional)}
}
\description{
Fit Cumulative Incidence Function curves
}
