% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{forestplot2}
\alias{forestplot2}
\title{Create a forest plot using ggplot2 (DEPRECATED)}
\usage{
forestplot2(
  model,
  conf.level = 0.95,
  orderByRisk = TRUE,
  colours = "default",
  showEst = TRUE,
  rmRef = FALSE,
  logScale = getOption("reportRmd.logScale", TRUE),
  nxTicks = 5
)
}
\arguments{
\item{model}{an object output from the glm or geeglm function, must be from a
logistic regression}

\item{conf.level}{controls the width of the confidence interval}

\item{orderByRisk}{logical, should the plot be ordered by risk}

\item{colours}{can specify colours for risks less than, 1 and greater than
1.0. Default is red, black, green}

\item{showEst}{logical, should the risks be displayed on the plot in text}

\item{rmRef}{logical, should the reference levels be removed for the plot?}

\item{logScale}{logical, should OR/RR be shown on log scale, defaults to
TRUE, or reportRmd.logScale if set. See https://doi.org/10.1093/aje/kwr156
for why you may prefer a linear scale.}

\item{nxTicks}{Number of tick marks supplied to the log_breaks function to
produce}
}
\value{
a plot object
}
\description{
#' @description
\strong{Deprecated}: Please use \code{\link[=forestplotMV]{forestplotMV()}} instead.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function will be removed in a future version.

This function will accept a log or logistic regression fit from glm or
geeglm, and display the OR or RR for each variable on the appropriate log
scale.
}
\keyword{plot}
