% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forestplotUV}
\alias{forestplotUV}
\title{Create a univariable forest plot using ggplot2}
\usage{
forestplotUV(
  response,
  covs,
  data,
  model = "glm",
  id = NULL,
  corstr = NULL,
  family = NULL,
  digits = getOption("reportRmd.digits", 2),
  conf.level = 0.95,
  colours = "default",
  showEst = TRUE,
  showRef = TRUE,
  logScale = getOption("reportRmd.logScale", TRUE),
  nxTicks = 5,
  showN = TRUE,
  showEvent = TRUE,
  xlim = NULL
)
}
\arguments{
\item{response}{character vector with names of columns to use for response}

\item{covs}{character vector with names of columns to use for covariates}

\item{data}{dataframe containing your data}

\item{model}{fitted model object (default "glm")}

\item{id}{character vector which identifies clusters. Only used for geeglm}

\item{corstr}{character string specifying the correlation structure. Only
used for geeglm}

\item{family}{description of the error distribution and link function to be
used in the model}

\item{digits}{number of digits to round to (default 2)}

\item{conf.level}{controls the width of the confidence interval (default 0.95)}

\item{colours}{can specify colours for risks less than, equal to, and greater
than 1.0. Default is green, black, red}

\item{showEst}{logical, should the risks be displayed on the plot in text?
Default is TRUE}

\item{showRef}{logical, should reference levels be shown? Default is TRUE}

\item{logScale}{logical, should OR/RR be shown on log scale? Defaults to TRUE}

\item{nxTicks}{Number of tick marks for x-axis (default 5)}

\item{showN}{Show number of observations per variable and category (default TRUE)}

\item{showEvent}{Show number of events per variable and category (default TRUE)}

\item{xlim}{numeric vector of length 2 specifying x-axis limits (ex c(0.2, 5))
Confidence intervals extending beyond these limits will be
shown with arrows.}
}
\value{
a ggplot object
}
\description{
This function creates forest plots from univariable regression models.
For new code, consider using forestplotMV() which can handle both adjusted
and unadjusted estimates.
}
\examples{
data("pembrolizumab")
forestplotUV(response = "orr",
             covs = c("change_ctdna_group", "sex", "age", "l_size"),
             data = pembrolizumab, family = 'binomial')
}
\keyword{plot}
