% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forestplotUVMV}
\alias{forestplotUVMV}
\title{Combine univariable and multivariable forest plot (DEPRECATED)}
\usage{
forestplotUVMV(UVmodel, MVmodel, ...)
}
\arguments{
\item{UVmodel}{an UV model object output from the forestplotUV function}

\item{MVmodel}{a MV model object output from the forestplotMV function}

\item{...}{additional arguments (ignored)}
}
\description{
This function is deprecated. Please use forestplotMV() with
include_unadjusted = TRUE instead.
}
