% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{formatp}
\alias{formatp}
\title{Specific p-value formatting}
\usage{
formatp(pvalues)
}
\arguments{
\item{pvalues}{a vector of p values}
}
\description{
If p < 0.001 returns "<0.001", if p < 0.01 returns p to 3 decimal places
otherwise returns p to 2 decimal places
}
\keyword{helper}
