% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geoR_boxcoxfit}
\alias{geoR_boxcoxfit}
\title{Parameter Estimation for the Box-Cox Transformation}
\usage{
geoR_boxcoxfit(object, xmat, lambda, lambda2 = NULL, add.to.data = 0)
}
\arguments{
\item{object}{a vector with the data}

\item{xmat}{a matrix with covariates values. Defaults to rep(1, length(y)).}

\item{lambda}{numerical value(s) for the transformation parameter lambda.
Used as the initial value in the function for parameter estimation. If not
provided default values are assumed. If multiple values are passed the one
with highest likelihood is used as initial value.}

\item{lambda2}{ogical or numerical value(s) of the additional transformation
(see DETAILS below). Defaults to NULL. If TRUE this parameter is also
estimated and the initial value is set to the absolute value of the minimum
data. A numerical value is provided it is used as the initial value.
Multiple values are allowed as for lambda.}

\item{add.to.data}{a constant value to be added to the data.}
}
\description{
This function is copied from the geoR package which has been removed from the
CRAN repository.
}
\details{
For more information see:
https://cran.r-project.org/web/packages/geoR/index.html
}
