% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblCode_orig.R
\name{getVariableLabels}
\alias{getVariableLabels}
\title{Show all variable label data frames}
\usage{
getVariableLabels()
}
\description{
Show all variable label data frames
}
\examples{
# add a separate table of data labels for the ctDNA data
data(ctDNA)
ctDNA_names <- data.frame(var=names(ctDNA),
label=c('Patient ID',
        'Study Cohort',
        'Change in ctDNA since baseline',
        'Number of weeks on treatment',
        'Percentage change in tumour measurement'))
setVariableLabels(ctDNA=ctDNA_names)

# Check that both the default and ctDNA label tables are set
getVariableLabels()
}
