% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hbld}
\alias{hbld}
\title{Bold strings for HTML output}
\usage{
hbld(strings)
}
\arguments{
\item{strings}{Vector of strings to bold}
}
\value{
Vector of strings wrapped in HTML bold span
}
\description{
Wraps strings in HTML bold formatting using inline CSS.
}
\keyword{helper}
