% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{lpvalue}
\alias{lpvalue}
\title{Formats p-values for LaTeX}
\usage{
lpvalue(x, sigdigits = 2)
}
\arguments{
\item{x}{an integer}

\item{sigdigits}{number of significant digit to report}
}
\description{
Returns <0.001 if pvalue is <0.001. Else rounds the pvalue to specified significant digits. Will bold the p-value if it is <= 0.05
}
