% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nicename}
\alias{nicename}
\title{lean strings for printing}
\usage{
nicename(strings, check_numbers = TRUE)
}
\arguments{
\item{strings}{vector of strings to give a nice name}

\item{check_numbers}{boolean indicating if numbers with decimals should be
checked for and retained.}
}
\description{
Returns strings with . and _ replaced by a space. This is nice when printing
column names of your dataframe in a report
}
\keyword{helper}
