% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{niceNum}
\alias{niceNum}
\title{Round retaining digits}
\usage{
niceNum(x, digits = 2)
}
\arguments{
\item{x}{a vector}

\item{digits}{numeric}
}
\description{
Round retaining digits
}
\keyword{helper}
