% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pembrolizumab}
\alias{pembrolizumab}
\title{Survival data

Survival status and ctDNA levels for patients receiving pembrolizumab}
\format{
A data frame with 94 rows and 15 variables:
\describe{
\item{id}{Patient ID}
\item{age}{Age at study entry}
\item{sex}{Patient Sex}
\item{cohort}{Study Cohort}
\item{l_size}{Target lesion size at baseline}
\item{pdl1}{PD L1 percent}
\item{tmb}{log of TMB}
\item{baseline_ctdna}{Baseline ctDNA}
\item{change_ctdna_group}{Did ctDNA increase or decrease from baseline to cycle 3}
\item{orr}{Objective Response}
\item{cbr}{Clinical Beneficial Response}
\item{os_status}{Overall survival status}
\item{os_time}{Overall survival time in months}
\item{pfs_status}{Progression free survival status}
\item{pfs_time}{Progression free survival time in months}
}
}
\source{
\url{https://www.nature.com/articles/s43018-020-0096-5}
}
\usage{
data('pembrolizumab')
}
\description{
Survival data

Survival status and ctDNA levels for patients receiving pembrolizumab
}
\keyword{datasets}
