% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{process_cif_medians}
\alias{process_cif_medians}
\title{Process CIF Median Values}
\usage{
process_cif_medians(
  fit,
  plot.event,
  stratalabs,
  median.lines = FALSE,
  median.text = FALSE,
  median.digits = 3,
  multiple_lines = FALSE
)
}
\arguments{
\item{fit}{CIF fit object}

\item{plot.event}{Events to plot}

\item{stratalabs}{Strata labels}

\item{median.lines}{Whether to calculate for median lines}

\item{median.text}{Whether to add median text}

\item{median.digits}{Number of digits for median}

\item{multiple_lines}{Whether there are multiple strata}
}
\value{
List with updated stratalabs and median values
}
\description{
Process CIF Median Values
}
