% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{process_covariate}
\alias{process_covariate}
\title{Process covariate variable (factor conversion, numeric cutoffs)}
\usage{
process_covariate(data, cov, cut = NULL, stratalabs = NULL)
}
\arguments{
\item{data}{Input data}

\item{cov}{Covariate column name}

\item{cut}{Numeric cutoff for continuous variables}

\item{stratalabs}{Custom strata labels}
}
\description{
Process covariate variable (factor conversion, numeric cutoffs)
}
