% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pstprn0}
\alias{pstprn0}
\title{Paste with parentheses (no spaces)}
\usage{
pstprn0(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Character string in format "x\href{x[2],x[3],...}{1}"
}
\description{
Similar to pstprn() but without spaces. Used for compact
formatting in survival plots.
}
\keyword{internal}
