% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pvalue}
\alias{pvalue}
\title{Formats p-values}
\usage{
pvalue(x, digits)
}
\arguments{
\item{x}{an integer}

\item{digits}{the number of significant digits to return}
}
\description{
Returns <0.001 if pvalue is <0.001. Else rounds the pvalue to specified
significant digits
}
