% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rmds}
\alias{rmds}
\title{Hacky awful work around to prevent errors when $ are used in covariate level names
only affects view output.}
\usage{
rmds(s)
}
\description{
Hacky awful work around to prevent errors when $ are used in covariate level names
only affects view output.
}
