% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_ggkmcif3_tests.r
\name{run_test}
\alias{run_test}
\title{Run a single test case}
\usage{
run_test(test_name, data, response, test_args = list())
}
\arguments{
\item{test_name}{Character string describing the test}

\item{data}{Data frame to use}

\item{response}{Character vector with time and status columns}

\item{test_args}{List of arguments to pass to ggkmcif3}
}
\description{
Run a single test case
}
