% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scrolling_table}
\alias{scrolling_table}
\title{Output a scrollable table}
\usage{
scrolling_table(knitrTable, pixelHeight = 500)
}
\arguments{
\item{knitrTable}{output from a call to knitr::kable or outTable}

\item{pixelHeight}{the height of the scroll box in pixels, default is 500}
}
\description{
This function accepts the output of a aa call to knitr::kable or
reportRmd::outTable and, if the output format is html, will produce a
scrollable table. Otherwise a regular table will be output for pandoc/latex
}
\examples{
data("pembrolizumab")
tab <- rm_covsum(data=pembrolizumab,maincov = 'change_ctdna_group',
covs=c('age','cohort','sex','pdl1','tmb','l_size'),full=FALSE)
scrolling_table(tab,pixelHeight=300)
}
