% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblCode.R
\name{set_var_labels}
\alias{set_var_labels}
\title{Set variable labels}
\usage{
set_var_labels(data, ...)
}
\arguments{
\item{data}{data frame containing variables to be labelled}

\item{...}{Name-label pairs the name gives the name of the column in the
output and the label is a character vector of length one.}
}
\description{
Set variable labels for a data frame using name-label pairs.
}
\details{
If no label is provided for a variable then the existing label will not be
changed. To remove a label set the label to NA.
}
\examples{
# set labels using name-label pairs
# and return labelled data frame
data("ctDNA")
ctDNA |> set_var_labels(
   ctdna_status="detectable ctDNA",
  cohort="A cohort label")
}
\seealso{
\code{\link[=set_labels]{set_labels()}} for setting variable labels using a data frame,
\code{\link[=extract_labels]{extract_labels()}} for creating a data frame of all variable labels,
\code{\link[=clear_labels]{clear_labels()}} for removing variable labels
}
