% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{setDataNames}
\alias{setDataNames}
\title{Set variable labels for data frames}
\usage{
setDataNames(.default, ...)
}
\arguments{
\item{...}{a sequence of named characters where the name specifies a data
frame containing data to analyse and the value is data frame containing
variable and label lookups.
@export}

\item{default}{A data frame containing the default labels for any variables
not specified in other tables}
}
\description{
To output use-friendly labels instead of variable names in the various
reportRmd functions you can supply data frames containing variable labels and names. A default
table can be supplied for all data within a report, or separate tables
can be provided for specific data frames.
}
