% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblCode_orig.R
\name{setVariableLabels}
\alias{setVariableLabels}
\title{Set variable labels}
\usage{
setVariableLabels(default, ...)
}
\arguments{
\item{default}{a data frame containing two columns, variable names and
variable labels to be used as default variable labels.}

\item{...}{pairs of data.frame, variable label pairs in the format
data=labels to provide variable labels specific to a data frame}
}
\description{
Assigns default or data frame specific tables containing variable labels
}
\details{
The variable labels need to be stored in two-column data frames. The first
column must contain the variable name and the second column the variable
label. The column names are not used.

If a data frame specific table of labels is provided it will be used for that
data frame only and labels contained in the default table will not be used.
If no labels are supplied for a variable then the variable name will be used
with '_' replaced by a space.
}
\examples{
# create a data frame contain variable labels and set as default
varLabels <- data.frame(vars=c('Var1','Var2','Var3'),lbls=c('Age','Sex','Score'))
setVariableLabels(varLabels)

# add a separate table of data labels for the ctDNA data
data(ctDNA)
ctDNA_names <- data.frame(var=names(ctDNA),
label=c('Patient ID',
        'Study Cohort',
        'Change in ctDNA since baseline',
        'Number of weeks on treatment',
        'Percentage change in tumour measurement'))
setVariableLabels(ctDNA=ctDNA_names)

# Check that both the default and ctDNA label tables are set
getVariableLabels()
}
