% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testData}
\alias{testData}
\title{Funky ctDNA data

There is a weird factor with all one level, the cohort variable contains a cohort level (and Cohort A) and for one cohort all the size changes are missing}
\format{
A data frame with 270 rows and 6 variables:
\describe{
\item{id}{Patient ID}
\item{cohort}{Study Cohort: A = Squamous cell carcinoma of soft pallate, B = Triple negative breast cancer, C = Ovarian, high grade serous, D = Melanoma, E = Other Solid Tumor, cohort -for testing only}
\item{badfactor}{for testing}
\item{ctdna_status}{Change in ctDNA since baseline}
\item{time}{Number of weeks on treatment}
\item{size_change}{Percentage change in tumour measurement}
}
}
\source{
\url{https://www.nature.com/articles/s43018-020-0096-5}
}
\usage{
testData
}
\description{
Funky ctDNA data

There is a weird factor with all one level, the cohort variable contains a cohort level (and Cohort A) and for one cohort all the size changes are missing
}
\keyword{datasets}
