% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uvmodels}
\alias{uvmodels}
\title{Aligning models and parameters}
\format{
A data frame with 16 rows and 5 variables:
\describe{
\item{type}{model type}
\item{family}{model linking family}
\item{gee}{boolean indicating if gee are used}
\item{autoreg_class}{class of model fit}
\item{beta}{description of output parameter}
}
}
\source{
internal
}
\usage{
uvmodels
}
\description{
Aligning models and parameters
}
\keyword{datasets}
