% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{uvsum}
\alias{uvsum}
\title{Get univariate summary dataframe}
\usage{
uvsum(..., markup, sanitize, forceWald)
}
\value{
Same as \code{uvsum2()}
}
\description{
Returns a dataframe corresponding to a univariate regression table
}
\details{
Univariate summaries for a number of covariates, the type of model can be
specified. If unspecified the function will guess the appropriate model based
on the response variable.

Confidence intervals are extracted using confint where possible. Otherwise
Student t distribution is used for linear models and the Normal distribution
is used for proportions.

returnModels can be used to return a list of the univariate models, which will
be the same length as covs. The data used to run each model will include all
cases with observations on the response and covariate. For gee models the data
are re-ordered so that the ids appear sequentially and proper estimates are
given.
}
\seealso{
\code{\link[=uvsum2]{uvsum2()}} for the current implementation
}
