% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{validate_and_prepare_data}
\alias{validate_and_prepare_data}
\title{Validate and prepare input data}
\usage{
validate_and_prepare_data(data, response, cov = NULL, print.n.missing = TRUE)
}
\arguments{
\item{data}{Input dataframe}

\item{response}{Character vector with time and status column names}

\item{cov}{Covariate column name (optional)}

\item{print.n.missing}{Whether to print missing data message}
}
\description{
Validate and prepare input data
}
