% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_compactsum.R, R/_refactored_compactsum.r
\name{xvar_function.default}
\alias{xvar_function.default}
\title{Helper function: xvar_function.default}
\usage{
\method{xvar_function}{default}(xvar, ...)

\method{xvar_function}{default}(xvar, ...)
}
\arguments{
\item{...}{}
}
\description{
Helper function: xvar_function.default

Default method for xvar_function
}
\keyword{internal}
