% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_refactored_compactsum.r
\name{xvar_function.rm_two_level}
\alias{xvar_function.rm_two_level}
\title{Two-level factor summarization}
\usage{
\method{xvar_function}{rm_two_level}(
  xvar,
  data,
  grp,
  covTitle = "",
  digits = 1,
  digits.cat = 0,
  iqr = TRUE,
  all.stats = FALSE,
  pvalue = TRUE,
  effSize = FALSE,
  show.tests = FALSE,
  percentage = "col"
)
}
\description{
Two-level factor summarization
}
\keyword{internal}
