% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{boxcoxfitRx}
\alias{boxcoxfitRx}
\title{fit box cox transformed linear model}
\usage{
boxcoxfitRx(f, data, lambda = F)
}
\arguments{
\item{f}{formula for the model. Currently the formula only works by using the name
of the column in a dataframe. It does not work by using $ or [] notation.}

\item{data}{dataframe containing data}

\item{lambda}{boolean indicating if you want to output the lamda used in the boxcox transformation. If so the function will return a list of length 2 with the model as the first element and a vector of length 2 as the second.}
}
\description{
Wrapper function to fit fine and gray competing risk model using function crr
from package cmprsk
}
\keyword{model}
