% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3.R
\name{create_cif_dataframe}
\alias{create_cif_dataframe}
\title{Create plotting dataframe for CIF curves}
\usage{
create_cif_dataframe(
  fit,
  stratalabs,
  plot.event = 1,
  conf.curves = FALSE,
  conf.type = "log",
  flip.CIF = FALSE
)
}
\arguments{
\item{fit}{CIF fit object}

\item{stratalabs}{Strata labels}

\item{plot.event}{Events to plot}

\item{conf.curves}{Whether to include confidence intervals}

\item{conf.type}{Confidence interval type}

\item{flip.CIF}{Whether to flip CIF (1-CIF)}
}
\description{
Create plotting dataframe for CIF curves
}
