% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{excelCol}
\alias{excelCol}
\title{Retrieve columns number from Excel columns specified as unquoted letters}
\usage{
excelCol(...)
}
\arguments{
\item{...}{unquoted excel column headers (i.e. excelCol(A,CG,AA)) separated by commas}
}
\description{
Retrieve columns number from Excel columns specified as unquoted letters
}
\examples{
## Find the column numbers for excel columns AB, CE and BB
# excelCol(AB,CE,bb)
}
