% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{extract_grays_test}
\alias{extract_grays_test}
\title{Extract Gray's test results from CIF fit}
\usage{
extract_grays_test(fit, plot.event = 1)
}
\arguments{
\item{fit}{CIF fit object or dataframe with test attribute}

\item{plot.event}{Events to test}
}
\description{
Extract Gray's test results from CIF fit
}
