% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmcif3_complete.R
\name{fit_cif_model}
\alias{fit_cif_model}
\title{Fit Competing Risks Model}
\usage{
fit_cif_model(data, response, cov = NULL)
}
\arguments{
\item{data}{Input dataframe}

\item{response}{Character vector with time and status column names}

\item{cov}{Covariate column name (optional)}
}
\value{
List containing fit object and group separator
}
\description{
Fit Competing Risks Model
}
