% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forestplot2}
\alias{forestplot2}
\title{Create a forest plot using ggplot2}
\usage{
forestplot2(
  model,
  conf.level = 0.95,
  orderByRisk = T,
  colours = "default",
  showEst = TRUE,
  rmRef = FALSE,
  logScale = TRUE,
  nxTicks = 5
)
}
\arguments{
\item{model}{an object output from the glm function, must be from a logistic
regression}

\item{conf.level}{controls the width of the confidence interval}

\item{orderByRisk}{logical, should the plot be ordered by risk}

\item{colours}{can specify colours for risks less than, 1 and greater than
1.0. Default is red, black, green}

\item{showEst}{logical, should the risks be displayed on the plot in text}

\item{rmRef}{logical, should the reference levels be removed for the plot?}

\item{logScale}{logical, should OR/RR be shown on log scale, defaults to TRUE}

\item{nxTicks}{Number of tick marks supplied to the log_breaks function to
produce}
}
\description{
This function will accept a log or logistic regression fit from glm, and
display the OR or RR for each variable on the appropriate log scale.
}
\examples{
glm_fit = glm(orr~change_ctdna_group+sex+age+l_size,
data=pembrolizumab,family = 'binomial')
forestplot2(glm_fit)
}
\keyword{plot}
