% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ggkmcif_paste}
\alias{ggkmcif_paste}
\title{combine components of a call to ggkmci}
\usage{
ggkmcif_paste(list_gg)
}
\arguments{
\item{list_gg}{A list of ggplot objects from \code{ggkmcif()}. (Deprecated)
Please use \code{\link[=ggkmcif2]{ggkmcif2()}} instead.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{ggkmcif()} was deprecated in version 0.1.2 and will be removed in a future version.
}
