% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nicename}
\alias{nicename}
\title{Clean strings for printing}
\usage{
nicename(strings)
}
\arguments{
\item{strings}{vector of strings to give a nice name}
}
\description{
Returns strings with . and _ replaced by a space. This is nice when printing column names of your dataframe in a report
}
\keyword{helper}
