% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{psthr}
\alias{psthr}
\title{Round and paste with parentheses}
\usage{
psthr(x, y = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{integer corresponding to the number of digits to round by}
}
\description{
Round and paste with parentheses
}
\keyword{helper}
