% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pstprn}
\alias{pstprn}
\title{Paste with parentheses}
\usage{
pstprn(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Paste with parentheses
}
\examples{
#pstprn(c(1,2,3,4,5))
#pstprn(c("Hello","Hi",2))
}
\keyword{helper}
