% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_cifsum}
\alias{rm_cifsum}
\title{Summarize cumulative incidence by group}
\usage{
rm_cifsum(
  data,
  time,
  status,
  group = NULL,
  eventcode = 1,
  cencode = 0,
  eventtimes,
  eventtimeunit,
  eventtimeLbls = NULL,
  CIwidth = 0.95,
  unformattedp = FALSE,
  na.action = "na.omit",
  showCounts = TRUE,
  showGraystest = TRUE,
  digits = 2,
  caption = NULL,
  tableOnly = FALSE,
  nocensored = FALSE
)
}
\arguments{
\item{data}{data frame containing survival data}

\item{time}{string indicating survival time variable}

\item{status}{string indicating the event status variable. The status
variable may be numeric, character, or factor and must identify the event
of interest and at least one competing event. Censoring and the event of
interest are specified using \code{cencode} and \code{eventcode}.}

\item{group}{string or character vector indicating the variable(s) to group
observations by. If this is left as NULL (the default), summaries are
provided for the entire cohort.}

\item{eventcode}{value of the status variable indicating the event of
interest. May be numeric, character, or factor-compatible. Default is 1.}

\item{cencode}{value of the status variable indicating censoring. May be
numeric, character, or factor-compatible. Default is 0.}

\item{eventtimes}{numeric vector specifying when event probabilities should
be calculated}

\item{eventtimeunit}{unit of time to suffix to the time column label if event
probabilities are requested, should be plural}

\item{eventtimeLbls}{if supplied, a vector the same length as eventtimes with
descriptions (useful for displaying years with data provided in months)}

\item{CIwidth}{width of the event probabilities, default is 95\%}

\item{unformattedp}{boolean indicating if you would like the p-value to be
returned unformatted (ie not rounded or prefixed with '<'). Should be used
in conjunction with the digits argument.}

\item{na.action}{default is to omit missing values, but can be set to throw
and error using na.action='na.fail'}

\item{showCounts}{boolean indicating if the event and total counts should be
displayed; default is TRUE}

\item{showGraystest}{boolean indicating Gray's test should be included in the
final table, default is TRUE}

\item{digits}{the number of digits to report in the event probabilities,
default is 2.}

\item{caption}{table caption for markdown output}

\item{tableOnly}{should a dataframe or a formatted object be returned}

\item{nocensored}{boolean indicating that the dataset genuinely contains no
censored observations. If the value specified by \code{cencode} is not
observed, the function will stop unless \code{nocensored = TRUE}; this
provides protection against an incorrectly specified censoring value.
Default is FALSE.}
}
\value{
A character vector of the event table source code, unless
tableOnly=TRUE in which case a data frame is returned
}
\description{
Displays event counts and event rates at specified time points for the entire
cohort and by group. Gray's test of differences in cumulative incidence is
displayed.
}
\examples{
library(survival)
data(pbc)

# Event probabilities at various time points with replacement time labels
rm_cifsum(data=pbc,time='time',status='status',
eventtimes=c(1825,3650),eventtimeLbls=c(5,10),eventtimeunit='yr')

# Event probabilities by one group
rm_cifsum(data=pbc,time='time',status='status',group='trt',
eventtimes=c(1825,3650),eventtimeunit='day')


# Event probabilities by multiple groups
rm_cifsum(data=pbc,time='time',status='status',group=c('trt','sex'),
eventtimes=c(1825,3650),eventtimeunit='day')

#'
# Character event status
pbc$status_chr <- factor(pbc$status, levels = c(0, 1, 2),
  labels = c("Censored", "Transplant", "Death"))

rm_cifsum(data = pbc, time = "time", status = "status_chr", group = "trt",
  eventcode = "Death", cencode = "Censored", eventtimes = c(1825, 3650),
  eventtimeLbls = c(5, 10), eventtimeunit = "yr")
}
