% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_mvsum2.R
\name{rm_mvsum}
\alias{rm_mvsum}
\title{Format a regression model nicely for 'Rmarkdown'}
\usage{
rm_mvsum(
  model,
  data,
  digits = getOption("reportRmd.digits", 2),
  covTitle = "",
  showN = TRUE,
  showEvent = TRUE,
  CIwidth = 0.95,
  vif = TRUE,
  whichp = c("levels", "global", "both"),
  caption = NULL,
  tableOnly = FALSE,
  p.adjust = "none",
  unformattedp = FALSE,
  nicenames = TRUE,
  include_unadjusted = FALSE,
  chunk_label,
  fontsize
)
}
\arguments{
\item{model}{model fit}

\item{data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} data is no longer required as
it is extracted from the model. This argument will be removed in the future}

\item{digits}{number of digits to round estimates to, does not affect p-values}

\item{covTitle}{character with the names of the covariate (predictor) column.
The default is to leave this empty for output or, for table only output to
use the column name 'Covariate'.}

\item{showN}{boolean indicating sample sizes should be shown for each
comparison, can be useful for interactions}

\item{showEvent}{boolean indicating if number of events should be shown. Only
available for logistic.}

\item{CIwidth}{width for confidence intervals, defaults to 0.95}

\item{vif}{boolean indicating if the variance inflation factor should be
included. See details}

\item{whichp}{string indicating whether you want to display p-values for
levels within categorical data ("levels"), global p values ("global"), or
both ("both"). Irrelevant for continuous predictors.}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{p.adjust}{p-adjustments to be performed. Uses the \link{p.adjust} function
from base R}

\item{unformattedp}{boolean indicating if you would like the p-value to be
returned unformatted (ie not rounded or prefixed with '<'). Should be used
in conjunction with the digits argument.}

\item{nicenames}{boolean indicating if you want to replace . and _ in strings
with a space}

\item{include_unadjusted}{Logical. If TRUE, includes univariate estimates
alongside multivariable estimates. Default is FALSE. The univariate models
are fit on the same complete-case sample as the multivariable model, so
that the two sets of estimates are directly comparable and the N column
applies to both. Any p-value adjustment requested via \code{p.adjust} is applied
to the multivariable p-values only. Covariates that appear only inside an
interaction term are appended to the end of the table with an unadjusted
estimate and no adjusted estimate.}

\item{chunk_label}{Deprecated, previously used in Word to allow
cross-referencing, this should now be done at the chunk level.}

\item{fontsize}{PDF/HTML output only, manually set the table fontsize}
}
\value{
A character vector of the table source code, unless tableOnly=TRUE in
which case a data frame is returned
}
\description{
Multivariable (or univariate) regression models are re-formatted for reporting
and a global p-value is added for the evaluation of factor variables.
}
\details{
Global p-values are likelihood ratio tests for lm, glm and polr models. For
lme models an attempt is made to re-fit the model using ML and if successful
LRT is used to obtain a global p-value. For lmer models (lme4), if the
lmerTest package is installed, Satterthwaite-based p-values and F-test global
p-values are used; otherwise Wald z-based p-values and chi-squared LRT global
p-values are returned. For glmer models (lme4), Wald z-based p-values are
used with chi-squared LRT global p-values. Estimates are exponentiated for
binomial (OR) and poisson/negative binomial (RR) families. For coxph models
the model is re-run without robust variances with and without each variable
and a LRT is presented. If unsuccessful a Wald p-value is returned. For GEE
and CRR models Wald global p-values are returned. For negative binomial
models a deviance test is used.

If the variance inflation factor is requested (VIF=TRUE, default) then a
generalised VIF will be calculated in the same manner as the car package.

As of version 0.1.1 if global p-values are requested they will be included in
the p-value column.

As of R 4.4.0 profile likelihood confidence intervals will
be calculated automatically and there is no longer an option to force Wald
tests.

The number of decimals places to display the statistics can be changed with
digits, but this will not change the display of p-values. If more significant
digits are required for p-values then use tableOnly=TRUE and format as
desired.
}
\examples{
data("pembrolizumab")
glm_fit = glm(change_ctdna_group~sex:age+baseline_ctdna+l_size,
data=pembrolizumab,family = 'binomial')
rm_mvsum(glm_fit)

#linear model with p-value adjustment
lm_fit=lm(baseline_ctdna~age+sex+l_size+tmb,data=pembrolizumab)
rm_mvsum(lm_fit,p.adjust = "bonferroni")
#Coxph
require(survival)
res.cox <- coxph(Surv(os_time, os_status) ~ sex+age+l_size+tmb, data = pembrolizumab)
rm_mvsum(res.cox, vif=TRUE)

# lmer (lme4 mixed effects model) - single random intercept
\donttest{
if (require(lme4)){
lmer_fit <- lme4::lmer(age ~ sex + pdl1 + (1|cohort), data = pembrolizumab)
rm_mvsum(lmer_fit)
}
}

# lmer with multiple random effects and global p-values
\donttest{
if (require(lme4) && require(geepack)){
data(dietox, package = "geepack")
dietox$Cu <- as.factor(dietox$Cu)
lmer_fit2 <- lme4::lmer(Weight ~ Cu + Time + (1|Pig) + (1|Litter), data = dietox)
rm_mvsum(lmer_fit2, whichp = "both")
}
}

# glmer (binomial mixed effects model) - odds ratios
\donttest{
if (require(lme4)){
data(cbpp, package = "lme4")
glmer_fit <- lme4::glmer(cbind(incidence, size - incidence) ~ period + (1|herd),
  data = cbpp, family = binomial)
rm_mvsum(glmer_fit)
}
}

# glmer.nb (negative binomial mixed effects model) - rate ratios
\donttest{
if (require(lme4) && require(geepack)){
data(dietox, package = "geepack")
dietox$Cu <- as.factor(dietox$Cu)
nb_fit <- lme4::glmer.nb(Weight ~ Cu + Time + (1|Pig), data = dietox)
rm_mvsum(nb_fit, whichp = "both")
}
}
}
\references{
John Fox & Georges Monette (1992) Generalized Collinearity
Diagnostics, Journal of the American Statistical Association, 87:417,
178-183, \doi{10.1080/01621459.1992.10475190}

John Fox and Sanford Weisberg (2019). An {R} Companion to Applied
Regression, Third Edition. Thousand Oaks CA: Sage.
}
