% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_mvsum}
\alias{rm_mvsum}
\title{Format a regression model nicely for Rmarkdown}
\usage{
rm_mvsum(
  model,
  data,
  digits = 2,
  covTitle = "",
  showN = FALSE,
  CIwidth = 0.95,
  caption = NULL,
  tableOnly = FALSE,
  p.adjust = "none",
  chunk_label,
  markup = T,
  sanitize = T,
  nicenames = T
)
}
\arguments{
\item{model}{model fit}

\item{data}{data that model was fit on (an attempt will be made to extract
this from the model)}

\item{digits}{number of digits to round to}

\item{covTitle}{character with the names of the covariate (predictor) column.
The default is to leave this empty for output or, for table only output to
use the column name 'Covariate'.}

\item{showN}{boolean indicating sample sizes should be shown for each
comparison, can be useful for interactions}

\item{CIwidth}{width for confidence intervals, defaults to 0.95}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{p.adjust}{p-adjustments to be performed (Global p-values only)}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{markup}{boolean indicating if you want latex markup}

\item{sanitize}{boolean indicating if you want to sanitize all strings to not
break LaTeX}

\item{nicenames}{boolean indicating if you want to replace . and _ in strings
with a space}
}
\description{
Multivariable (or univariate) regression models are re-formatted for
reporting and a global p-value is added for the evaluation of factor
variables.
}
\details{
Global p-values are likelihood ratio tests for lm, glm and polr models. For
lme models an attempt is made to re-fit the model using ML and if,successful
LRT is used to obtain a global p-value. For coxph models the model is re-run
without robust variances with and without each variable and a LRT is
presented. If unsuccessful a Wald p-value is returned. For GEE and CRR models
Wald global p-values are returned.
}
\examples{
glm_fit = glm(change_ctdna_group~sex:age+baseline_ctdna+l_size,
data=pembrolizumab,family = 'binomial')
rm_mvsum(glm_fit)
}
