% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_uvsum}
\alias{rm_uvsum}
\title{Output several univariate models nicely in a single table}
\usage{
rm_uvsum(
  response,
  covs,
  data,
  digits = 2,
  covTitle = "",
  caption = NULL,
  tableOnly = FALSE,
  removeInf = TRUE,
  p.adjust = "none",
  chunk_label,
  id = NULL,
  corstr = NULL,
  family = NULL,
  type = NULL,
  strata = 1,
  nicenames = TRUE,
  testing = FALSE,
  showN = TRUE,
  CIwidth = 0.95,
  reflevel = NULL
)
}
\arguments{
\item{response}{string vector with name of response}

\item{covs}{character vector with the names of columns to fit univariate
models to}

\item{data}{dataframe containing data}

\item{digits}{number of digits to round to}

\item{covTitle}{character with the names of the covariate (predictor) column.
The default is to leave this empty for output or, for table only output to
use the column name 'Covariate'.}

\item{caption}{character containing table caption (default is no caption)}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{removeInf}{boolean indicating if infinite estimates should be removed
from the table}

\item{p.adjust}{p-adjustments to be performed (Global p-values only)}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{id}{character vector which identifies clusters. Only used for geeglm}

\item{corstr}{character string specifying the correlation structure. Only
used for geeglm. The following are permitted: '"independence"',
'"exchangeable"', '"ar1"', '"unstructured"' and '"userdefined"'}

\item{family}{description of the error distribution and link function to be
used in the model. Only used for geeglm}

\item{type}{string indicating the type of univariate model to fit. The
function will try and guess what type you want based on your response. If
you want to override this you can manually specify the type. Options
include "linear", "logistic", "coxph", "crr", "boxcox", "ordinal", "geeglm"}

\item{strata}{character vector of covariates to stratify by. Only used for
coxph and crr}

\item{nicenames}{boolean indicating if you want to replace . and _ in strings
with a space}

\item{testing}{boolean to indicate if you want to print out the covariates
before the model fits.}

\item{showN}{boolean indicating if you want to show sample sizes}

\item{CIwidth}{width of confidence interval, default is 0.95}

\item{reflevel}{manual specification of the reference level. Only used for
ordinal regression This will allow you to see which model is not fitting if
the function throws an error}
}
\description{
A table with the model parameters from running separate univariate models on
each covariate. For factors with more than two levels a Global p-value is
returned.
}
\details{
Global p-values are likelihood ratio tests for lm, glm and polr models. For
lme models an attempt is made to re-fit the model using ML and if,successful
LRT is used to obtain a global p-value. For coxph models the model is re-run
without robust variances with and without each variable and a LRT is
presented. If unsuccessful a Wald p-value is returned. For GEE and CRR models
Wald global p-values are returned.
}
\examples{
# Coxph model with 90\% CI
rm_uvsum(response = c('os_time','os_status'),
covs=c('age','sex','baseline_ctdna','l_size','change_ctdna_group'),
data=pembrolizumab,CIwidth=.9)
# Linear model with default 95\% CI
rm_uvsum(response = 'baseline_ctdna',
covs=c('age','sex','l_size','pdl1','tmb'),
data=pembrolizumab)
}
\seealso{
\code{\link{uvsum}},\code{\link{lm}},\code{\link{glm}},\code{\link{crr}},
\code{\link{coxph}},
\code{\link{lme}},\code{\link{geeglm}},\code{\link{polr}}
}
