% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rmds}
\alias{rmds}
\title{Replace dollar signs with html for proper HTML output}
\usage{
rmds(s)
}
\arguments{
\item{s}{a character vector}
}
\description{
Replace dollar signs with html for proper HTML output
}
\keyword{helper}
