% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{uvsum}
\alias{uvsum}
\title{Get univariate summary dataframe}
\usage{
uvsum(
  response,
  covs,
  data,
  digits = 2,
  id = NULL,
  corstr = NULL,
  family = NULL,
  type = NULL,
  strata = 1,
  markup = TRUE,
  sanitize = TRUE,
  nicenames = TRUE,
  testing = FALSE,
  showN = TRUE,
  CIwidth = 0.95,
  reflevel = NULL
)
}
\arguments{
\item{response}{string vector with name of response}

\item{covs}{character vector with the names of columns to fit univariate
models to}

\item{data}{dataframe containing data}

\item{digits}{number of digits to round to}

\item{id}{character vector which identifies clusters. Only used for geeglm}

\item{corstr}{character string specifying the correlation structure. Only
used for geeglm. The following are permitted: '"independence"',
'"exchangeable"', '"ar1"', '"unstructured"' and '"userdefined"'}

\item{family}{description of the error distribution and link function to be
used in the model. Only used for geeglm}

\item{type}{string indicating he type of univariate model to fit. The
function will try and guess what type you want based on your response. If
you want to override this you can manually specify the type. Options
include "linear", "logistic", "coxph", "crr", "boxcox", "ordinal", "geeglm"}

\item{strata}{character vector of covariates to stratify by. Only used for
coxph and crr}

\item{markup}{boolean indicating if you want latex markup}

\item{sanitize}{boolean indicating if you want to sanitize all strings to not
break LaTeX}

\item{nicenames}{boolean indicating if you want to replace . and _ in strings
with a space}

\item{testing}{boolean to indicate if you want to print out the covariates
before the model fits.}

\item{showN}{boolean indicating if you want to show sample sizes}

\item{CIwidth}{width of confidence interval, default is 0.95}

\item{reflevel}{manual specification of the reference level. Only used for
ordinal This will allow you to see which model is not fitting if the
function throws an error}
}
\description{
Returns a dataframe corresponding to a univariate regression table
}
\details{
Univariate summaries for a number of covariates, the type of model can be
specified. If unspecified the function will guess the appropriate model based
on the response variable.
}
\seealso{
\code{\link{lm}},\code{\link{glm}},\code{\link{crr}},\code{\link{coxph}},
\code{\link{lme}},\code{\link{geeglm}},\code{\link{polr}}
}
\keyword{dataframe}
